# openapi.json
{"openapi":"3.0.0","paths":{"/api/status":{"get":{"operationId":"status","parameters":[],"responses":{"200":{"description":"Service status","content":{"application/json":{"schema":{"$ref":"#/components/schemas/StatusResponse"},"example":{"status":"ok","database":"ok"}}}},"429":{"description":"Rate limit exceeded"}},"summary":"Health check","tags":["Status"]}},"/api/channels":{"get":{"operationId":"getChannels","parameters":[{"name":"order","required":false,"in":"query","description":"Sort direction","schema":{"enum":["desc","asc"],"type":"string"}},{"name":"sort","required":false,"in":"query","description":"Field to sort by","schema":{"enum":["displayName","xmlId"],"type":"string"}},{"name":"limit","required":false,"in":"query","description":"Items per page (max 100, default 20)","schema":{"example":20,"type":"number"}},{"name":"page","required":false,"in":"query","description":"Page number (min 1)","schema":{"example":1,"type":"number"}}],"responses":{"200":{"description":"Paginated list of channels","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PaginatedChannelsResponse"}}}},"429":{"description":"Rate limit exceeded"}},"summary":"List channels","tags":["Channels"]}},"/api/channels/search":{"get":{"operationId":"searchChannels","parameters":[{"name":"limit","required":false,"in":"query","description":"Items per page (max 100, default 20)","schema":{"example":20,"type":"number"}},{"name":"page","required":false,"in":"query","description":"Page number (min 1)","schema":{"example":1,"type":"number"}},{"name":"q","required":true,"in":"query","description":"Search term","schema":{"type":"string"}}],"responses":{"200":{"description":"Paginated matching channels","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PaginatedChannelsResponse"}}}},"429":{"description":"Rate limit exceeded"}},"summary":"Search channels by name","tags":["Channels"]}},"/api/channels/tnt":{"get":{"operationId":"getTntChannels","parameters":[],"responses":{"200":{"description":"TNT channels in order","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ChannelWithCurrent"}}}}},"429":{"description":"Rate limit exceeded"}},"summary":"List TNT free channels in broadcast order","tags":["Channels"]}},"/api/channel/{id}":{"get":{"operationId":"getChannelDetails","parameters":[{"name":"id","required":true,"in":"path","description":"Channel UUID or xmlId","schema":{"type":"string"}},{"name":"day","required":false,"in":"query","description":"DD/MM/YYYY (Europe/Paris)","schema":{"example":"20/05/2026","type":"string"}}],"responses":{"200":{"description":"Channel with current and daily programs","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChannelDetailsResponse"}}}},"404":{"description":"Channel not found"},"429":{"description":"Rate limit exceeded"}},"summary":"Get a channel with its current and daily programs","tags":["Channels"]}},"/api/program/{id}":{"get":{"description":"Lookup a program by its UUID.","operationId":"getProgram","parameters":[{"name":"id","required":true,"in":"path","description":"Program UUID","schema":{"type":"string"}}],"responses":{"200":{"description":"Program details","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProgramWithChannel"}}}},"404":{"description":"Program not found"},"429":{"description":"Rate limit exceeded"}},"summary":"Get a program by id","tags":["Programs"]}},"/api/programs/now":{"get":{"description":"Returns a paginated list of programs airing now.","operationId":"getCurrentPrograms","parameters":[{"name":"order","required":false,"in":"query","description":"Sort direction","schema":{"enum":["desc","asc"],"type":"string"}},{"name":"sort","required":false,"in":"query","description":"Field to sort by","schema":{"enum":["startAt","title"],"type":"string"}},{"name":"limit","required":false,"in":"query","description":"Items per page (max 100, default 20)","schema":{"example":20,"type":"number"}},{"name":"page","required":false,"in":"query","description":"Page number (min 1)","schema":{"example":1,"type":"number"}}],"responses":{"200":{"description":"Paginated list of current programs","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PaginatedProgramsResponse"}}}},"429":{"description":"Rate limit exceeded"}},"summary":"List currently airing programs","tags":["Programs"]}},"/api/programs/{day}":{"get":{"description":"Returns a paginated list of programs for the specified day.","operationId":"getProgramsByDay","parameters":[{"name":"day","required":true,"in":"path","description":"Date in YYYY-MM-DD format","schema":{"type":"string"}},{"name":"limit","required":false,"in":"query","description":"Items per page (max 100, default 20)","schema":{"example":20,"type":"number"}},{"name":"page","required":false,"in":"query","description":"Page number (min 1)","schema":{"example":1,"type":"number"}}],"responses":{"200":{"description":"Paginated list of programs for the day","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PaginatedProgramsResponse"}}}},"400":{"description":"Invalid date format"},"429":{"description":"Rate limit exceeded"}},"summary":"List programs for a specific day","tags":["Programs"]}},"/api/tmdb/init":{"get":{"operationId":"handleNewPrograms","parameters":[],"responses":{"200":{"description":"Initialization triggered"}},"summary":"Initialize TMDB details for all programs","tags":["TMDB"]}},"/api/tmdb/sync":{"get":{"operationId":"syncProgramScores","parameters":[{"name":"title","required":false,"in":"query","description":"Filter by program title","schema":{"type":"string"}}],"responses":{"200":{"description":"Sync completed"}},"summary":"Sync TMDB scores for current programs","tags":["TMDB"]}},"/api/searxng/sync":{"get":{"operationId":"syncPosters","parameters":[{"name":"title","required":false,"in":"query","description":"Filter by program title","schema":{"type":"string"}}],"responses":{"200":{"description":"Sync triggered"}},"summary":"Sync poster images from SearXNG for programs without poster","tags":["SearXNG"]}},"/api/xml-tv/run":{"get":{"description":"Require x-forwarded-for header with valid value to start cron","operationId":"startJob","parameters":[{"name":"x-forwarded-for","in":"header","description":"Header to prevent running cron","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":""}},"summary":"Start cront to download new xml and update database","tags":["xml-tv"]}}},"info":{"title":"TV FR API","description":"API for French TV channels and programs based on racacax/XML-TV-Fr","version":"1.0","contact":{}},"tags":[],"servers":[{"url":"https://tv-api.miceli.click"}],"components":{"schemas":{"StatusResponse":{"type":"object","properties":{"status":{"type":"string"},"database":{"type":"string"}},"required":["status","database"]},"TmdbDetails":{"type":"object","properties":{"id":{"type":"string"},"title":{"type":"string"},"mediaType":{"type":"string","nullable":true,"enum":["tv","movie"]},"tmdbId":{"type":"number","nullable":true},"originalName":{"type":"string","nullable":true},"popularity":{"type":"number","nullable":true},"voteCount":{"type":"number","nullable":true},"poster":{"type":"string","nullable":true},"secondaryPoster":{"type":"string","nullable":true},"searxngSearchedAt":{"format":"date-time","type":"string","nullable":true},"tmdbSyncAt":{"format":"date-time","type":"string","nullable":true},"tmdbUrl":{"type":"string","nullable":true}},"required":["id","title","mediaType","tmdbId","originalName","popularity","voteCount","poster","secondaryPoster","searxngSearchedAt","tmdbSyncAt","tmdbUrl"]},"ProgramSummary":{"type":"object","properties":{"id":{"type":"string"},"title":{"type":"string"},"subTitle":{"type":"string","nullable":true},"desc":{"type":"string","nullable":true},"xmlStart":{"type":"string"},"xmlStop":{"type":"string"},"startAt":{"format":"date-time","type":"string"},"stopAt":{"format":"date-time","type":"string"},"ratingIcon":{"type":"string","nullable":true},"icon":{"type":"string","nullable":true},"episode":{"type":"string","nullable":true},"rating":{"type":"string","nullable":true},"categories":{"type":"array","items":{"type":"string"}},"credits":{"type":"array","items":{"type":"string"}},"channelXmlId":{"type":"string"},"details":{"nullable":true,"type":"object","allOf":[{"$ref":"#/components/schemas/TmdbDetails"}]}},"required":["id","title","subTitle","desc","xmlStart","xmlStop","startAt","stopAt","ratingIcon","icon","episode","rating","categories","credits","channelXmlId","details"]},"ChannelWithCurrentAndUrls":{"type":"object","properties":{"urls":{"type":"array","items":{"type":"string"}},"current":{"nullable":true,"type":"object","allOf":[{"$ref":"#/components/schemas/ProgramSummary"}]},"id":{"type":"string"},"displayName":{"type":"string"},"xmlId":{"type":"string"},"icon":{"type":"string","nullable":true}},"required":["urls","current","id","displayName","xmlId","icon"]},"PaginatedChannelsResponse":{"type":"object","properties":{"channels":{"type":"array","items":{"$ref":"#/components/schemas/ChannelWithCurrentAndUrls"}},"total":{"type":"number"},"totalPages":{"type":"number"},"count":{"type":"number"},"limit":{"type":"number"}},"required":["channels","total","totalPages","count","limit"]},"ChannelWithCurrent":{"type":"object","properties":{"current":{"nullable":true,"type":"object","allOf":[{"$ref":"#/components/schemas/ProgramSummary"}]},"id":{"type":"string"},"displayName":{"type":"string"},"xmlId":{"type":"string"},"icon":{"type":"string","nullable":true}},"required":["current","id","displayName","xmlId","icon"]},"ChannelSummary":{"type":"object","properties":{"id":{"type":"string"},"displayName":{"type":"string"},"xmlId":{"type":"string"},"icon":{"type":"string","nullable":true}},"required":["id","displayName","xmlId","icon"]},"ChannelDetailsResponse":{"type":"object","properties":{"channel":{"$ref":"#/components/schemas/ChannelSummary"},"currentProgram":{"nullable":true,"type":"object","allOf":[{"$ref":"#/components/schemas/ProgramSummary"}]},"dayPrograms":{"type":"array","items":{"$ref":"#/components/schemas/ProgramSummary"}}},"required":["channel","currentProgram","dayPrograms"]},"ProgramWithChannel":{"type":"object","properties":{"id":{"type":"string"},"title":{"type":"string"},"subTitle":{"type":"string","nullable":true},"desc":{"type":"string","nullable":true},"xmlStart":{"type":"string"},"xmlStop":{"type":"string"},"startAt":{"format":"date-time","type":"string"},"stopAt":{"format":"date-time","type":"string"},"ratingIcon":{"type":"string","nullable":true},"icon":{"type":"string","nullable":true},"episode":{"type":"string","nullable":true},"rating":{"type":"string","nullable":true},"categories":{"type":"array","items":{"type":"string"}},"credits":{"type":"array","items":{"type":"string"}},"channelXmlId":{"type":"string"},"details":{"nullable":true,"type":"object","allOf":[{"$ref":"#/components/schemas/TmdbDetails"}]},"channel":{"$ref":"#/components/schemas/ChannelSummary"}},"required":["id","title","subTitle","desc","xmlStart","xmlStop","startAt","stopAt","ratingIcon","icon","episode","rating","categories","credits","channelXmlId","details","channel"]},"ProgramDetails":{"type":"object","properties":{"id":{"type":"string"},"title":{"type":"string"},"subTitle":{"type":"string","nullable":true},"desc":{"type":"string","nullable":true},"xmlStart":{"type":"string"},"xmlStop":{"type":"string"},"startAt":{"format":"date-time","type":"string"},"stopAt":{"format":"date-time","type":"string"},"ratingIcon":{"type":"string","nullable":true},"icon":{"type":"string","nullable":true},"episode":{"type":"string","nullable":true},"rating":{"type":"string","nullable":true},"categories":{"type":"array","items":{"type":"string"}},"credits":{"type":"array","items":{"type":"string"}},"channelXmlId":{"type":"string"},"details":{"nullable":true,"type":"object","allOf":[{"$ref":"#/components/schemas/TmdbDetails"}]},"channelDisplayName":{"type":"string"}},"required":["id","title","subTitle","desc","xmlStart","xmlStop","startAt","stopAt","ratingIcon","icon","episode","rating","categories","credits","channelXmlId","details","channelDisplayName"]},"PaginatedProgramsResponse":{"type":"object","properties":{"programs":{"type":"array","items":{"$ref":"#/components/schemas/ProgramDetails"}},"total":{"type":"number"},"totalPages":{"type":"number"},"count":{"type":"number"},"limit":{"type":"number"}},"required":["programs","total","totalPages","count","limit"]}}}}